# 03-Azure/01-04-AI/01_Agentic_SDLC/src/api-ts/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "summary": "Returns products with optional search and filters",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Case-insensitive text search across product name, SKU, and description"
                    },
                    {
                        "in": "query",
                        "name": "supplierId",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Filter products by supplier ID"
                    },
                    {
                        "in": "query",
                        "name": "minPrice",
                        "schema": {
                            "type": "number",
                            "minimum": 0
                        },
                        "description": "Minimum product price"
                    },
                    {
                        "in": "query",
                        "name": "maxPrice",
                        "schema": {
                            "type": "number",
                            "minimum": 0
                        },
                        "description": "Maximum product price"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filter parameter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "VALIDATION_ERROR"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Validation error: minPrice cannot be greater than maxPrice"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new product",
                "tags": [
                    "Products"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "summary": "Get a product by ID",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "summary": "Update a product",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a product",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "type": "object",
                "required": [
                    "productId",
                    "supplierId",
                    "name",
                    "price",
                    "sku",
                    "unit"
                ],
                "properties": {
                    "productId": {
                        "type": "integer",
                        "description": "The unique identifier for the product"
                    },
                    "supplierId": {
                        "type": "integer",
                        "description": "The ID of the supplier providing this product"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the product"
                    },
                    "description": {
                        "type": "string",
                        "description": "Detailed description of the product"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "description": "The current price of the product"
                    },
                    "sku": {
                        "type": "string",
                        "description": "Stock keeping unit code for the product"
                    },
                    "unit": {
                        "type": "string",
                        "description": "Unit of measure for the product (e.g., \"piece\")"
                    },
                    "imgName": {
                        "type": "string",
                        "description": "Filename of the product image"
                    },
                    "discount": {
                        "type": "number",
                        "format": "float",
                        "description": "Discount percentage (if applicable) expressed as a decimal (e.g., 0.25 for 25%)"
                    }
                }
            },
...
        }
    },
...
}